# app/openapi.yaml
openapi: 3.0.0
info:
  title: Petnet API's Doc
  description: This file content all the routes on Petnet application.
  version: 1.0.0
servers:
  - url: "https://petnet-8e037ce96247.herokuapp.com"
    description: Main production server.
  - url: "htpp://localhost:5000"
    description: Internal staging server for development.
...
paths:
...
  /patients/register:
    post:
      tags:
        - Petnet-Patient
      summary: Register a patient
      description: Registers a new patient with the provided information.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                  description: The nickname of the patient.
                age:
                  type: integer
                  description: The age of the patient.
                weight:
                  type: number
                  description: The weight of the patient.
                owner_id:
                  type: integer
                  description: The ID of the owner of the patient.
                profile_photo:
                  type: string
                  format: binary
                  description: The profile photo of the patient (base64 encoded image).
      security:
        - ApiKeyAuth: []
      responses:
        '201':
          description: Patient created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating successful patient creation.
                  patient:
                    $ref: '#/components/schemas/Patient'
        '400':
          description: Bad Request. Missing required fields or profile photo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure due to missing required fields or profile photo.
        '500':
          description: Internal Server Error. Error creating patient.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure during patient creation.
                  error:
                    type: string
                    description: Additional information about the error.
...
  /patients/search:
    get:
      tags:
        - Petnet-Patient
      summary: Search for patients
      description: Searches for patients based on a query string. Use  '*' to get all the patients registered.
      parameters:
        - in: query
          name: s
          required: true
          description: The search query string.
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Patients found successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Patient'
        '400':
          description: Bad Request. Missing search query.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure due to missing search query.
        '404':
          description: Not Found. Patients not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure due to inability to find patients.
  /patients/update/information/{patient_id}:
    put:
      tags:
        - Petnet-Patient
      summary: Update patient information
      description: Updates the information of a specific patient.
      parameters:
        - in: path
          name: patient_id
          required: true
          description: The ID of the patient.
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                  description: The nickname of the patient.
                age:
                  type: integer
                  description: The age of the patient.
                weight:
                  type: number
                  description: The weight of the patient.
              required:
                - nickname
                - age
                - weight
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Patient information updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating successful update.
                  patient:
                    $ref: "#/components/schemas/Patient"
        '400':
          description: Bad Request. Missing required fields.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure due to missing required fields.
        '404':
          description: Not Found. Patient not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure due to patient not found.
        '500':
          description: Internal Server Error. Error updating patient.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure during patient update.
                  error:
                    type: string
                    description: Additional information about the error.
...
  /patients/delete/{patient_id}:
    delete:
      tags:
        - Petnet-Patient
      summary: Delete a patient
      description: Deletes a specific patient.
      parameters:
        - in: path
          name: patient_id
          required: true
          description: The ID of the patient to delete.
          schema:
            type: integer
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Patient deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating successful deletion.
        '500':
          description: Internal Server Error. Error deleting patient.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure during deletion.
                  error:
                    type: string
                    description: Additional information about the error.
...
  /owner/register:
    post:
      tags:
        - Petnet-Owner
      summary: Register an owner
      description: Registers a new owner with the provided information.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  description: The first name of the owner.
                last_name:
                  type: string
                  description: The last name of the owner.
                email:
                  type: string
                  format: email
                  description: The email address of the owner.
                phone_number:
                  type: string
                  description: The phone number of the owner.
                document:
                  type: string
                  description: The document number of the owner.
              required:
                - first_name
                - last_name
                - email
      security:
        - ApiKeyAuth: []
      responses:
        '201':
          description: Owner registered successfully.
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Owner'
                  
        '400':
          description: Bad Request. Invalid input data or email already registered.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating the reason for the failure.
        '500':
          description: Internal Server Error. Error registering the owner.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure during registration.
                  error:
                    type: string
                    description: Additional information about the error.    
  /owner/search:
    get:
      tags:
        - Petnet-Owner
      summary: Search for owners
      description: Searches for owners based on a query string. Use  '*' to get all the owners registered.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: s
          in: query
          description: The search query string.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Owners found successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Owner'                
        '400':
          description: Bad Request. The query parameter is missing.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating that the query parameter is missing.
        '404':
          description: Not Found. No owners found matching the search query.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating that no owners were found matching the search query.                    
  /owner/update/information/{owner_id}:
    put:
      tags:
        - Petnet-Owner
      summary: Update owner information
      description: Updates the information of an owner with the provided owner_id.
      parameters:
        - in: path
          name: owner_id
          required: true
          description: The ID of the owner to update.
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  description: The updated first name of the owner.
                last_name:
                  type: string
                  description: The updated last name of the owner.
                email:
                  type: string
                  format: email
                  description: The updated email address of the owner.
                phone_number:
                  type: string
                  description: The updated phone number of the owner.
                document:
                  type: string
                  description: The updated document number of the owner.
              required:
                - first_name
                - last_name
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Owner information updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating the success of the update operation.
                  owner:
                    $ref: '#/components/schemas/Owner'
                    
        '400':
          description: Bad Request. Invalid input data or email already registered.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating the reason for the failure.
        '404':
          description: Owner not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating that the owner was not found.
        '500':
          description: Internal Server Error. Error updating owner information.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure during the update operation.
                  error:
                    type: string
                    description: Additional information about the error.       
  /owner/get/information/{owner_id}:
    get:
      tags:
        - Petnet-Owner
      summary: Get owner information by ID
      parameters:
        - in: path
          name: owner_id
          required: true
          schema:
            type: integer
          description: The ID of the owner to retrieve.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating the success of the operation.
                  owner:
                    $ref: '#/components/schemas/Owner'
        '404':
          description: Owner not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating that the owner was not found.
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating that there was an error getting owner information.
                  error:
                    type: string
                    description: A description of the error.                    
  /owner/delete/{owner_id}:
    delete:
      tags:
        - Petnet-Owner
      summary: Delete an owner
      description: Deletes the owner with the provided owner_id.
      parameters:
        - in: path
          name: owner_id
          required: true
          description: The ID of the owner to delete.
          schema:
            type: integer
            format: int64
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Owner deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating the success of the deletion operation.
        '404':
          description: Owner not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating that the owner was not found.
        '500':
          description: Internal Server Error. Error deleting the owner.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating failure during the deletion operation.
                  error:
                    type: string
                    description: Additional information about the error.   
components:
  schemas:
...
    Patient:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier of the patient.
        nickname:
          type: string
          description: The nickname of the patient.
        age:
          type: integer
          description: The age of the patient.
        date_of_register:
          type: string
          format: date-time
          description: The date of registration of the patient.
        weight:
          type: number
          description: The weight of the patient.
        profile_photo:
          type: string
          format: uri
          description: The URL of the profile photo of the patient.
        user_id:
          type: integer
          description: The ID of the user associated with the patient.
        owner_id:
          type: integer
          description: The ID of the owner associated with the patient.
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
          description: List of photos associated with the patient.
    Photo:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier of the photo.
        photo:
          type: string
          format: uri
          description: The URL of the photo.
        filename:
          type: string
          description: The filename of the photo.
        date_of_register:
          type: string
          format: date-time
          description: The date of registration of the photo.
        patient_id:
          type: integer
          description: The ID of the patient associated with the photo.
        description:
          type: string
          description: Description of the photo.
        probability:
          type: number
          description: Probability associated with the photo.
        predicted_class:
          type: string
          description: Predicted class associated with the photo. 
    Owner:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier of the owner.
        first_name:
          type: string
          description: The first name of the owner.
        last_name:
          type: string
          description: The last name of the owner.
        email:
          type: string
          format: email
          description: The email address of the owner.
        phone_number:
          type: string
          description: The phone number of the owner.
        document:
          type: string
          description: The document information of the owner.
        patients:
          type: array
          items:
            $ref: '#/components/schemas/Patient'
          description: List of patients associated with the owner.
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization 
security:
  - ApiKeyAuth: []